+++ thoth/common/exceptions.py
@@ -31,3 +31,11 @@ class ConfigurationError(ThothCommonException):
 
 class WorkflowError(ThothCommonException):
     """Raised on workflow related issues."""
+
+
+class NotKnownThothIntegration(ThothCommonException):
+    """An exception raised if the given integration is not known to Thoth."""
+
+
+class QebHwtInputsMissing(ThothCommonException):
+    """An exception raised if there are inputs missing for QebHwt Github App."""